Move msg conversion to convert module
